Relocate scm info to GH plugin
